+++ src/game/Board.h
@@ -16,7 +16,7 @@ public:
 
     int getRows();
     int getCols();
-    std::vector<std::vector<float>> getBoard();
+    std::vector<std::vector<int>> getBoard();
 
     // Singleton based on:
     // https://codereview.stackexchange.com/a/173935
@@ -32,7 +32,7 @@ private:
     Board();
     int cols;
     int rows;
-    std::vector<std::vector<float>> board;
+    std::vector<std::vector<int>> board;
 };
